requirements.txt
einops==0.3.0
natsort==8.3.1
numpy==1.21.5
opencv_contrib_python==4.2.0.32
Pillow==9.2.0
piq==0.7.0
skimage==0.0
tabulate==0.8.10
torch==1.12.1
torchvision==0.13.1